+++ core-ng/src/main/java/core/framework/internal/module/ModuleContext.java
@@ -56,6 +56,7 @@ public class ModuleContext {
         route(HTTPMethod.GET, "/_sys/vm", (LambdaController) diagnosticController::vm, true);
         route(HTTPMethod.GET, "/_sys/thread", (LambdaController) diagnosticController::thread, true);
         route(HTTPMethod.GET, "/_sys/heap", (LambdaController) diagnosticController::heap, true);
+        route(HTTPMethod.GET, "/_sys/proc", (LambdaController) diagnosticController::proc, true);
         route(HTTPMethod.GET, "/_sys/property", new PropertyController(propertyManager), true);
     }