+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 7.6.4 (11/19/2020 - )
+* kafka: update snappy-java to 1.1.8.1, (according to release note, there is small performance improvement)
 
 ### 7.6.3 (11/16/2020 - 11/19/2020)
 * module: DBConfig.repository() returns repository object, to make it easier to extend, same applied to es search type and mongo collection