+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * kafka: update snappy-java to 1.1.8.1, (according to release note, there is small performance improvement)
 * kafka: added kafka().maxRequestSize() to configure max request size, this setting must be consistent with broker
 * stat: track and warn on java process high vmRSS usage, for docker OOMKill cases better enable -XX:NativeMemoryTracking=summary with detailed usage report
+* diagnostic: added /_sys/proc to report process info (mainly memory usage of java process) on linux
 
 ### 7.6.3 (11/16/2020 - 11/19/2020)
 * module: DBConfig.repository() returns repository object, to make it easier to extend, same applied to es search type and mongo collection